pyproject.toml
[tool.poetry]
name = "hinglish-voice-bot"
version = "0.1.0"
description = "Voice bot for blue-collar candidate screening in Hinglish"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]
package-mode = true

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
websockets = "^12.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
speechrecognition = "^3.8.1"
pydub = "^0.25.1"
google-cloud-speech = "^2.21.0"
google-cloud-texttospeech = "^2.16.3"
openai = "^1.3.0"
azure-cognitiveservices-speech = "^1.34.0"
numpy = "^1.24.0"
scipy = "^1.11.0"
httpx = "^0.25.0"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 